Add a macro to define or delete copy constructors and assignment operators
Summary:
As a rule of thumb, we should always explictly define copy constructors and assignment operators (sometimes called the "rule of 5").
The problem is that this is overly verbose, since it's 5 lines of code that are always the same.
This diff adds two macros that can be used to define or delete those.
I have also added a default constructor since it doesn't do any harm, and is also part of the rule of 5.
There are a few cases where this shouldn't be used: for instance, when defining a base class used for polymorphism (e.g, `Kind`, `MethodConstraint`, etc.) since the destructor needs to be `virtual`.

Reviewed By: anwesht

Differential Revision: D43415469

fbshipit-source-id: 41dca8b38d7aeb1064d11c09731b9789e399cac3