rename TypeNameConstraint -> TypePatternConstraint
Summary: Context: Plan to re-use the "TypeNameConstraint" name to be *exact* name matches instead of regex matches when implementing signature_match. Since the pattern constraint has been the only way to match names, we need to disambiguate the two. The next diff will implement a `TypeNameConstraint` that does exact matches.

Reviewed By: arthaud

Differential Revision: D43400784

fbshipit-source-id: fb123ce221127919d752236249671b02869bf1f1