Refactor of the forward analysis for invoke
Summary:
This is a non functional change, including some minor tweaks of the forward analysis to make it more readable.
This started while working on a backward analysis, realizing it was pretty hard to follow the logic on invoke, especially for multi-source rules. This diff improves slightly the state of the code.

A non-exhaustive list of changes:
* Disambiguate function overloads: we have many `check_flows` functions with slightly different parameters. It makes it really confusing. We could use better naming here.
* Avoid optional arguments: in general I think we should avoid those.
* Avoid `if (ptr)`, prefer `if (ptr != nullptr)`: This is less ambiguous and clearly shows that the variable is a pointer.

Reviewed By: anwesht

Differential Revision: D43520217

fbshipit-source-id: 5e3926d98d8fe1d39d3efec1b77b0ad5d941a12d