Allow shim on anonymous overrides of methods
Summary:
When the shimmed-method has a parameter override associated with it, we create a new instance of `Method` to wrap the original `DexMethod` with the associated parameter-overrides.
This diff updates the shims lookup to check for the original method instead. The shim-target method will still be the updated method with the parameter-overrides.

Reviewed By: anwesht

Differential Revision: D42713844

fbshipit-source-id: aeac2cc3e127a26fbe17f9082f24e90059f4f842