Add a method to append to output paths
Summary:
When we read a specific path in a backward taint tree, we will need to add path elements to the `output_paths` tree for the special "local" taints used to infer propagation (`LocalReturn` and `LocalArgument`). This is necessary to properly infer output paths.
This diff implements that method. It will be used in follow-ups.
Note that since we currently use artificial sources, we do the opposite: we add paths to input paths, cf the `append_to_artificial_source_input_paths` method.

Reviewed By: anwesht

Differential Revision: D43582067

fbshipit-source-id: 917a5f293ac75ffb597fd47a5909b7e1bafa0017