Release the memory used for the alias analysis [backward analysis 15/n]
Summary:
We have seen a big increase in memory usage since the addition of the forward alias analysis.
I suspect that the problem is that we keep the whole alias analysis fixpoint in memory while computing the forward taint analysis.
This diff fixes that problem by using scopes. We could also move those into functions.

Reviewed By: anwesht

Differential Revision: D43578925

fbshipit-source-id: 89f5e702aa85d7e417737e799a784b29141657d5