Add a filter method on taint
Summary:
We currently have a `map` method on most of the taint representation, except `Taint`. This diff fixes that.
Then, this diff implements a `filter` method that only keeps frames that pass the given predicate.

Note that one might think this could be done with a map, for instance:
```
taint.map([](Frame& frame) {
  if (predicate(frame)) {
    frame.set_to_bottom();
  }
});
```
This doesn't work since it breaks an invariant of the taint representation (see `GroupHashedSetAbstractDomain::map`). It is also less idiomatic than a proper `filter` method.

Reviewed By: anwesht

Differential Revision: D43579629

fbshipit-source-id: 0185b00375f7a2384556e2117d7ab49b2b23a111