Fix nitpicks and improve consistency
Summary:
This is a non functional change. I went through a lot of the codebase while working on the backward analysis.
This diff fixes a few nitpicks I found and improves consistency of the codebase.
This includes: using `const` when possible, fix confusing naming, remove trivial comments, remove unused includes.

Reviewed By: anwesht

Differential Revision: D43582308

fbshipit-source-id: eaa057e6fb2beca41b604947357a62dedb34e002