Add a data structure to store triggered partial sinks to propagate up [backward analysis 16/n]
Summary:
# Context
We are working on implementing a backward analysis in Mariana Trench.
This analysis will be done after the forward alias analysis and forward taint analysis, and will be responsible for propagating sinks up and inferring propagations. Once this is done, we will remove entirely the use of artificial sources.

# Problem
Multi-source rules requires special handling in the forward and backward analysis. Whenever a partial sink is fulfilled, the counterpart should be propagated up. This is currently done directly in the forward analysis by inferring sinks using artificial sources. Once we migrated to a backward analysis, we will need to propagate those sinks up in the backward analysis.

# Solution
This diff implements a class `FulfilledPartialKindResults` which will be responsible for holding the sinks discovered in the forward analysis, that must be propagated up in the backward analysis. It is basically a map from invoke instruction to the fulfilled state. It needs an extra map for artificial callees too.

Note that this is currently unused. See follow-up diffs.

Reviewed By: anwesht

Differential Revision: D43580115

fbshipit-source-id: 794da5238891577e527c8e0d18c426f5ea21d386