populate method mappings when building the call graph
Summary:
We had a fun bug where refactoring model generators to use the `signature_match` syntax caused fewer models to get generated, as the call graph generation happens *after* we create method mappings.

Expose the API in MethodMappings to create mappings for a given method, and ensure we create methods properly as we create new methods in the call graph.

Note: I was initially hoping to pass in MethodMappings references to the ::create() API and force these to exist, but it turns out that we have `Method::from_json()` APIs which also create these methods. I don't think it's great to pass in the MethodMappings object to something that should be a pure function, so ended up going with a separate call after we create methods in call graph building.

Reviewed By: anwesht

Differential Revision: D43515147

fbshipit-source-id: 947fc24259cb4c432bd777da406a880a9223be17