refactor `IntentDataFeatureGenerator` to use signature_match instead of signature_pattern
Summary:
One interesting thing to note here: I changed a model to fire on all Intent <init>'s instead of filtering ones with the Uri type in them. Reasoning's that the models are only firing on parameters that have the appropriate type, so the search on signature match should be pretty fast regardless.

0.40s -> 0.03s runtime for FBLite.

Reviewed By: anwesht

Differential Revision: D43539609

fbshipit-source-id: eb64a285db60bc30d08306d3a70d7944020805da