Do not add features to argument on taint that was propagated
Summary:
We currently add features to argument after applying propagations.
I believe we should NOT do this.
If a user-declared propagation want to add a feature, it should be marked as such. In general, user-declared annotations should not affect other user-declared annotations.
For instance, specifying `attach-to-sink` on a model which also has a user-declared sink should NOT add the features on the sink.
This is for the sake of simplicity and performance.

Reviewed By: anwesht

Differential Revision: D43589789

fbshipit-source-id: 7c54f7cdb5a5eda0ca1015957f9609cdec2af44e