add "method" information to "origin" in callee port frames
Summary:
Context: We'd like to expose the actual names of the callees in origins as well - SAPP displays richer leaf names in other tools, and we'd like to do the same in Mariana Trench.

This diff adds the singular `method` field to each origin, if there is one. A future diff will amend SAPP parsing to use these as the leaf names.

Reviewed By: anwesht

Differential Revision: D43839650

fbshipit-source-id: 0e53de13ba3d5a1ba1ed7fafa7d880c1db03bb9d