Remove type matching requirement on shims
Summary:
As highlighted in D42678105 we have a problem knowing the type of the BroadcastManager used to register a specific ActionReceiver. D42678105 solves part of this problem by allowing us to taint the injection field and carry this over to addActionReceiver.

This diff allows us to take the taint on `Argument(0) addActionReceiver` and transfer it to `Argument(2) onReceive` where the class type matches `Argument(2) addActionReceiver`. Speaking more clearly it effectively allows us to transfer the taint of the injected type to the Intent which is the actual untrusted data source.

To do this however I had to shim non-matching types which is why Im removing the error in Shim.cpp. I do think this is ok to do as we deliberately here want the flow to make sense for the end user (security engineer) and the behaviour is limited to just doing this on addActionReceiver. Maybe we make a JSON parameter that deliberately has to be specified to bypass this error to make it more obvious. An alternative is a more elaborate ArtificialMethod but that felt worse than doing this.

I also hardcoded disable_parameter_type_overrides == true to get around a FN  when shimming + parameter overrides interact T143329376.

Reviewed By: anwesht

Differential Revision: D41302226

fbshipit-source-id: 07acd0b70f8e9df7b061ff1b7c69467e2ad8646b