add an `--enable-cross-component-analysis` option to analysis
Summary:
Context:

While working on the implementation of cross-component analysis, we'll need to add a bunch of edges to the call graph and extra information to the taint representation. I'd like to avoid performance hits to regular Mariana Trench runs until we've proven out that the feature is good to go.

Reviewed By: anwesht

Differential Revision: D44025725

fbshipit-source-id: b5f93c94cfaaa81c275a1d6b8cadb3c504f6b92a