abstract shim access to be per-caller via a Shims class
Summary:
Context:

The ulterior motive for this change is cross-component analysis. For cross-component analysis, we'll need per-caller connections, as we'll want to mark `startActivity` as flowing to different methods that have `getIntent()` calls depending on which classes were specified in the intent prior to calling `startActivity`.

This diff sets up the environment for that change, and centralizes the logic for fetching shims to not allow unfiltered access. In the future, we'll want to add per-caller shims (probably a Method -> Method -> Shim map) to make the call graph consistent for taint analysis.

Note that we'll need a richer representation of the connections than what shims can offer, this bit is only for getting the dependencies right within the fixpoint.

Reviewed By: anwesht

Differential Revision: D43949196

fbshipit-source-id: 17052e169938578f1a742fdb9dd6e48a700b20ce