hide MethodToShimMap within the Shims class
Summary: The idea is to make the `Shims` class and more contained APIs the only way to interact with shims. Required a bit of restructuring how we report errors, and having things take a Shims object in instead of creating separate maps and merging them, but it's a relatively simple change.

Reviewed By: arthaud

Differential Revision: D44032949

fbshipit-source-id: f9b6b21b4a8c38b78059967c0b36ef712ac8e529