Build a new model from scratch at each iteration
Summary:
# Problem
When performing an iteration and analyzing a method, we add any inferred sources/sinks/propagations directly onto the previous model. This is fine until we call `model.approximate()` which can lead to collapsing taint.
The problem is that this could lead to collapsing user defined taint, which is probably not expected.

# Solution
We should instead create a new empty model and add inferred sources/sinks/propagations to that one, then call `model.approximate()`. In the end, the global fixpoint will join the previous model and the new one anyway.

Reviewed By: anwesht

Differential Revision: D44311527

fbshipit-source-id: fd485fdfe770a76e4d7a0e6c0cedf9f6c9c01b9e