introduce pass to compute cross-component hooks in methods
Summary:
Overall idea: For cross-component analysis, we're aiming to capture intents routed to other classes in multiple forms. One of the primitives we need for this, which this diff implements the structure of, is a pre-taint fixpoint analysis that collects all potential classes an intent routes to.

The analysis is barebones for now - we only look for the intent constructor that takes a context and a class, use the results of our reflection analysis to try and figure out which class this came from, and have a per-method set of potential routed intent classes (stored as `DexType *`'s).

Looking for feedback on:

- Analysis structure: Not sure I'm using sparta right.
- Data structure we're using for the analysis - would folks prefer a map from instruction -> DexType* set being returned from the get go? For now the implementation would cheat, but this is probably what we want longer term.

Next up:

- The second primitive that collects all `getIntent()`'s for classes, and constructs a `DexType* -> [Method]` mapping, that'll allow us to update our call graph.
- Understanding of the different forms of specifying classes for cross-component analysis (via extras, setComponent, etc.).
- Augmenting shims to use this analysis' results

Reviewed By: yuhshin-oss

Differential Revision: D44234931

fbshipit-source-id: cd9b256007445c2ff54e43e14751ad532c32732e