Approximate override models
Summary:
We use a set of heuristics to prevent Mariana Trench from blowing up. For instance, we limit the number of leaves in the access path tree for sources, sinks and propagations. This is mostly done by the `Model::approximate` call.

I recently noticed that we do not approximate override models. Override models are the join of all models for overriding methods of a given method in a base class. Override models can become quite big for methods with a lot of overrides. Let's try to approximate those models, which should give us performance benefits at the cost of some false positives.

Reviewed By: yuhshin-oss

Differential Revision: D44374286

fbshipit-source-id: 6d906f85340fec3b53f30fe7a0d9ebc377606005