Implement abstract tree shaping
Summary:
This diff adds a method on `AbstractTreeDomain` that can be used to re-shape a tree without losing any elements.
Any branches from the tree that is not present in the mold is collapsed into its parents.

This will be used in a follow-up to prevent blowing up on recursive methods, such as the test introduced previously.

Reviewed By: anwesht

Differential Revision: D44500888

fbshipit-source-id: 5fd4b5d8ca80c7d44021a281f5034656706558e0