Add integration tests with recursive sinks
Summary:
This test shows what happens with recursive methods with sinks.
On `recursive_parent` we infer sinks on `this`, `parent`, `parent.parent`, etc.
On `recursive_praent_x` we infer sinks on `x`, `parent.x`, `parent.parent.x`, etc.
This is unoptimal since those sinks are then propagated. This is detrimental to performance.

In a follow-up, we will implement abstract tree shaping and molding to get rid of these duplicate sinks.

Reviewed By: anwesht

Differential Revision: D44496749

fbshipit-source-id: 7246c3f4fc715c7a867ea6def08d3b526c4b2fe2