build intent routing data in shims
Summary:
This diff implements an intermediate data structure intended to be used for shim generation later. The idea is to build 2 different mappings: Class -> set of methods that call getIntent(), and method -> set of different component classes this method may route to.

We gate the construction of the data structure behind the enable_cross_component_analysis option, so our regular analyses shouldn't be affected.

Had to expose a bunch of guts publically to facilitate testing. Let me know if there's better ways of doing this in C++.

Reviewed By: anwesht

Differential Revision: D44433401

fbshipit-source-id: cfe7a8ed7c4b41a56f4e63b07c4f86ea45ef0784