Implement a backward analysis [backward analysis 17/17]
Summary:
This implements a backward analysis for mariana trench, which is responsible for propagating sinks up and inferring propagations.
This works by performing a fixpoint on the reversed control flow graph and analyzing instructions backwards.

We use special taint kinds to infer propagations to `this` (using `LocalArgument(0)`) and to the return value (using `LocalReturn`). These are attached to the right memory locations at the beginning of the backward analysis and propagated up like regular sinks.

This replaces the usage of artificial sources in the forward analysis and can provide better precision in specific cases (see integration tests), although it can also lead to false positives in some cases.

In the future, this will also make it easier to infer a propagation collapse depth, which will avoid always collapsing when propagating.

This could also provide performance improvements, since we are tracking less taint using this strategy: only return values and `ParameterMemoryLocation(0)` are tainted with `LocalArgument(0)` and `LocalReturn`, while the approach using artificial sources required to taint every single parameter and propagate taint down.

Reviewed By: anwesht

Differential Revision: D43562469

fbshipit-source-id: 959cacb14ee315e95c2784ee85a8f36929b9ce6e