Make factories singletons
Summary:
We have a few factory classes (`KindFactory`, `FeatureFactory` and `TransformsFactory`) responsible for returning unique instances and managing the memory of elements used in the taint representation.
Those factories are independent of the analyzed code or taint configuration. They do not have any logic. They could potentially be re-used for different analyses.
Let's make them singletons (i.e, global variables), so we only ever create one of these.

This will be useful in the future, when we want to create features out of thin air, without having access to the context because of restrictive APIs.

Reviewed By: anwesht

Differential Revision: D44669615

fbshipit-source-id: 31d8455394b45bcb209101d07cc8b22103d84fb4