add a method to check whether callers exist in the call graph without constructing intermediate objects
Summary:
Context: Saw that the Interprocedural.cpp file was using the callees() and artificial_callees() of a method to schedule re-analysis. Not only will this lead to boilerplate with the intent routing callees coming up, we also create a bunch of unnecessary vectors when the use only intends to check for non-emptiness.

Let's clean this up.

Reviewed By: anwesht

Differential Revision: D44629576

fbshipit-source-id: 5c3e687c0fceed8ed08885fe196bf1ae9269f159