Add a feature when collapsing taint during widening
Summary:
We limit the height of taint trees when applying widening, which happens when the analysis gets stuck on a loop.
To limit the height, we simply collapse taint trees at a depth of 4.
This diff adds a feature `via-widen-broadening` to taint that was collapsed during that operation.

This diff also adds necessary scaffolding for the tito collapse depth work, which will require to specify the behavior on certain abstract tree operations.

Reviewed By: sinancepel

Differential Revision: D44708516

fbshipit-source-id: dc9d687f535c0380157817a5914668cc8b9866dd