Avoid using `std::function` with a non-const reference parameter
Summary:
We have many methods in our codebase taking a `std::function` as a parameter, such as `map` methods.
Some of those take a `std::function<void(T&)>`, and except it to mutate a given `T`.
Others use `std::function<T(const T&)>` or `std::function<T(T)>` which expects it to return the computed/updated value.

This is for instance what `sparta::PatriciaTreeMap` uses.

Let's make it uniform across the codebase. For the reasons below, we are chosing `std::function<T(T)>` as the best API:
* This is already used by `sparta::PatriciaTreeMap`, which we probably don't want to change.
* `std::function<void(T&)>` can be error-prone since one could write a function returning `T`, which will be silently converted to `void(T&)`.
* This is more common in a "functional" programming language: function take values and return new ones.
* `void(T&)` can have a small performance advantage for data structures that work best with mutations (e.g, hash maps). I believe that `T(T)` is still a good compromise, assuming that the function will almost-always mutate the object. `T(T)` can be optimized/inlined easily be the compiler and benefit from move-semantic.

Reviewed By: sinancepel

Differential Revision: D44747691

fbshipit-source-id: 929abd11c6c3636c8b1d6211297a7bb132a8a9c3