Use template parameters instead of `std::function`
Summary:
Using `std::function` can add a small overhead, since `std::function` can lead to memory allocations (it is implemented using type erasure) and virtual dispatching (call to a virtual method). It also makes inlining practically impossible for the compiler.

Instead, let's use template parameters.

Function as template parameters have a few downsides on their own. The code is slightly more verbose, the interface of functions can be a bit obscure and template errors are hard to debug. I believe this is worth it for the performance win, for very common methods (like `map` and `filter`).

Reviewed By: dkgi, yuhshin-oss

Differential Revision: D44751897

fbshipit-source-id: 274787a4a4235ea05549462e70c03cd4faec438c