Multi-source features: Use "inferred_features" (previously "locally_inferred_features")
Summary:
We are planning on moving `locally_inferred_features` into `CalleePortFrames`. All frames that share a (callee, position, callee port) tuple should have the same set of locally inferred features. This invariant should hold even today, though it is not stored that way.

In multi-source rules, when a partial sink (label:x) is triggered on argument(2) as follows:

```
void triggeredSink(Object object) {
  MultiSource.sink(object, /* partial sink label:x*/ sourcesAandB());
}
```

A triggered sink is created for argument(1) for each rule that contains a partial sink with label:x. The features from `sourcesAandB()` become part of that sink. Today, these features are stored in "locally_inferred_features". This violates the invariant because source kind A may have different features from source kind B. Specifically:

```
Callee = MultiSource.sink
Position=line1 of function
Callee port=argument(1))

KindA -> Features:{source-a}
KindB -> Features:{source-b}
```

The fix is to store them as just "inferred_features" that are unique to the kind. The downside is that we lose the ability to show that these "local features" were introduced in the frame in the UI, but they aren't really "local" anyway. The features were introduced somewhere in `sourcesAandB()` but we just don't show that subtrace today.

**More documentation about locally inferred features**
This refers to features accumulated within the method during the analysis. For example:

```
void f() {
  int arg = 1;
  /* various operations on arg** ... */

  // Frame at sink:
  // Callee = StaticMethods.to_sink
  // Call position = line no. of to_sink call
  // Callee port = argument(0)
  // The locally_inferred_features accumulated above should be the same,
  // regardless of what kinds `arg` contains.
  StaticMethods.to_sink(arg);
}
```

Reviewed By: sinancepel

Differential Revision: D44733378

fbshipit-source-id: 053278355bdd74648ccd462e51b41ea500fa45ae