Add hooks to the abstract tree domain on sink and hoist (1/n tito collapse depth)
Summary:
# Context
The current analysis always collapses taint trees when applying propagations (except with the `no-collapse-on-propagation` mode). This can lead to a lot of false positives, since we would then consider a whole object as tainted instead of having a single field tainted.

For instance:
```
class Class {}

Class identity(Class x) {
  return x;
}
void issue() {
  Class x = new Class();
  x.foo = source(); // x.foo is tainted
  Class y = identity(x); // y (whole object) is tainted because of collapsing!
  sink(y.bar); // False positive, since y.bar is an attribute of a tainted object.
}
```

In the following stack, we implement what we refer to as "tito collapse depth" or propagation collapse depth. For each propagation, we will infer an integer that corresponds to the maximum depth of the input taint tree. A depth of 0 means the input taint should be entirely collapsed, which is the current behavior.

In the example above, we could easily infer that `identity` does not need to collapse it's input taint since it preserves the structure of `x`. This would prevent us from finding an issue, since only `y.foo` will be tainted and not `y.bar`.

# In this diff
This diff adds a few hooks to the abstract tree domain, which ared used to implement the output path tree for propagations. These hooks will be necessary to infer a sound collapse depth.

Reviewed By: anwesht

Differential Revision: D45045953

fbshipit-source-id: 42e973b09facd41e4bb6ce23b64eeaee49c6fc1a