Infer a propagation collapse depth for propagations (3/n tito collapse depth)
Summary:
See context in D45045953.

This diff implements the inference of the collapse depth of propagations.
This should improve precision by avoiding unnecessary taint broadening.

This is inferred in the backward analysis alongside propagation. The initial collapse depth is set to 4 (reasonable default) and then propagated up. Tree `read`s append to the output path while reducing the collapse depth.

To make this sound, we need to reset the collapse depth to 0 (meaning, collapse everything) when the backward taint is collapsed itself. This happens when tracking too many fields, for instance (see test `approximate_return_access_path`).

Reviewed By: anwesht

Differential Revision: D45121235

fbshipit-source-id: ad9fbaded53a332c7af9a5ac75ca8206d8d7e75d