add `extends` variant of signature_match queries
Summary:
Chose to include both single and multiple parents to be accepted for this for consistency with the parents/names variants. Added `include_self` in the obvious manner (defaults to true, can be overridden). This is definitely getting to the limit of complexity we'd be wanting to support in the `signature_match` shorthand, so if people want to disallow extending multiple classes or disallow not including self in the easy form, I'm ok with that decision as well.

Lastly, changed the json validation to count the parent- and name- related members in the JSON, and error if it's not exactly one. This allows us to avoid erroring in multiple places and decouples erroring and parsing logic a bit.

Reviewed By: GerbenJavado

Differential Revision: D45219597

fbshipit-source-id: 93a823361de2d4fd3b608ecca81f3af538fc8877