Stop emitting local features/positions within Frame's json
Summary:
As the title says.

Also, user features, which were previously reported as local features, are now merged with inferred features instead. These features can be kind-specific and should stay within the `Frame`. This also means that they will not be shown in the sapp UI. Note that the UI also did not handle these features correctly before. It does not take into account the "kind" that they apply to, so showing user features there could be misleading.

Reviewed By: anwesht

Differential Revision: D44981085

fbshipit-source-id: 2c1331f630efc89dd23a85f85fb25f08811da1e2