Backward Analysis: Support call-effect propagation
Summary:
Propagating data stored in the Intent via `startActivity()` and`this.getIntent()` calls require call effects and propagation.

The "Intent" call effect is a special port for taint to flow via. Propagations need to support it for taint to flow through:

```
// Some other Activity
public void issue() {
  Intent = new Intent(context, MyActivity.class);
  intent.putExtra("session", source());
  // Invisible call edge to MyActivity.m()
  startActivity(intent);
}

// In MyActivity:
public class MyActivity extends Activity {
  public void m() {
    // The intent is tainted in issue() above.
    // Indicated by a propagation from CallEffectIntent -> Return
    // on this.getIntent()
    Intent intent = this.getIntent();
    sink(intent);
  }
}
```

In this diff, CallEffect ports are only allowed as the input path of the propagation (not output path which needs to be return/argument).

The way propagation happens is the same as it would on an Argument port, except that the taint is stored in the `Model::call_effect_sinks_` rather than tainting the respective register for the argument.

____

Reviewed By: arthaud

Differential Revision: D45429046

fbshipit-source-id: 177310d0775d40a910e9496e6229c7c61a3a2d8c