Forward analysis: Flow sources via startActivity shim into CallEffectIntent
Summary:
With the following already in place:

* Shims: `startActivity(intent)` -> relevant methods that call `getIntent()`
* Inferring of CallEffectIntent sinks on methods that call `getIntent()` if the intent flows into a sink.

This change updates the forward analysis to handle the following case:

```
public void issue(Context context) {
  Intent intent = new Intent(context, TargetActivity.class);
  intent.putExtra("session", source());

  // Shims to [TargetActivity.<methodsThatGetIntent>]
  // Internally, shim has mapping:
  //   port:CallEffectIntent -> register_of:intent
  // This change updates forward analysis to process this mapping
  // and forwards the taint, creating issues where applicable.
  startActivity(intent);
}
```

**Other Considerations**

There's a lot of copy-pasta going on where the differences are between:

* Reading regular sinks vs. call-effect sinks.
* Reading parameter taint vs port taint.
* Reading artificial callees vs intent routing callees (also `ArtificialCallee`s).

Could put them in a bunch of lambdas but that's kinda abstract and hurts readability. I'm guessing we can merge a couple of underlying data structures and reduce the number of `check_*_flows()` altogether. Will look into this once I'm done with the next step (propagating CallEffectIntent sinks in backward analysis).

Reviewed By: anwesht

Differential Revision: D45594403

fbshipit-source-id: 5bfdaeef02bce394af199215b7afd39c71ccecb7