Artificial Callee: Map port to register
Summary:
Similar to the changes in shims, this maps the target callee's port to the corresponding register in the caller for `ArtificialCallee`s.

This mapping is only populated for regular shims and intentionally ignored for life-cycle and reflection shims. Specifying of ports-to-parameters is currently needed only for `shims.intent_routing_targets()`.

Reviewed By: anwesht

Differential Revision: D45594402

fbshipit-source-id: 8463c0f23d4e742b86f41bfd4d38ea56602c1036