Always print model generator parse errors
Summary:
We are currently logging parse errors for model generators at log level 3, which means it won't show up by default. That seems like a pretty confusing behaviour, but is related to the fact that we try to parse all json files (include rule definitions, lifecycle, etc.) as model generators.

To improve the currently state, let's print all model generator parse errors. To detect files that are not model generators, we silently skip json files that are not objects (e.g, `{...}`). This is pretty hacky but is a good alternative until we introduce proper file extensions.

Reviewed By: GerbenJavado

Differential Revision: D46640293

fbshipit-source-id: 0097d3f420fee2895e9171651f7648138d2d0cb6