CalleePortFrames: Remove use of GroupHashedSetAbstractDomain
Summary:
Prior to this change, in CalleePortFrames, frames were stored in a map:
```
Kind -> GroupHashedSet<Frame>

hash key =
  frame.callee + frame.callee_port + frame.call_position + frame.kind
```

Recall that the Taint structure is:

```
Taint
  -> CalleeFrames
  -> CallPositionFrames
  -> CalleePortFrames
  -> Kind
  -> GroupHashedSet<Frame>
```

The key for the group hashed set is already captured in the Taint structure, which means the set size is always 1.

In that case, it can simply be stored as:
```
Kind -> Frame
```
This change achieves that.

Reviewed By: anwesht

Differential Revision: D46789065

fbshipit-source-id: 880fac8854e53d99672d93ae81132d02da63aadb