Adding option for making analysis flow-sensitivity...
Summary:
... taking into account that invoke-*, check-cast, and instance-of instructions simplify when the target class isn't instantiable. The basic idea, and the code rewriting to reflect the simplified behavior, is directly liftedn from the RemoveUninstantiablesPass (which will be disabled later in the stack).

Here's an example from the new test suite where before running RMU and RemoveInstantiablesPass would get stuck:
```
public static void root() { clone(null); }

static DataHolder clone(DataHolder data_holder) {
    Data data = data_holder.get_data();
    return new DataHolder(data);
}
```
RMU wouldn't be able to remove the `DataHolder` constructor, `get_data` method, or the `DataHolder` class itself, as they are all referenced by hold reachable from a root. RemoveUninstantiablePass wouldn't be able to do anything, as `DataHolder` has a constructor and an instance is being created, so it's not uninstantiable.

However, with this new option in RMU, we'll only incrementally visit the control-flow paths of a method, and pausing visiting when we come across a field access or invocation to a member of a class that hasn't been identified as "instantiable" yet. Here, we'll never get to the `DataHolder` constructor in that way, because we keep being paused at the invocation to the instance method `get_data`.

The new option is disabled by default, which then behaves as if all instruction target classes are instantiable, which is the current behavior. So by itself this is a behavior-preserving diff.

Reviewed By: jimmycFB

Differential Revision: D46418308

fbshipit-source-id: 8c48c5439f9ecc88985855f1908426c141aae7bc