Add a model generator name class
Summary:
This diff introduces a class that represents the unique name of a model generator.
This is useful for logging - for instance, it generates `<name>:<index>` for json model generators.
It will also be used in a follow-up to store in models the set of model generators that affected them.

Reviewed By: anwesht

Differential Revision: D47245637

fbshipit-source-id: 33454fdf4a993e29d216667eee56e243b1c53747