Do not prevent taint collapsing on methods returning this
Summary:
In the past, we used the `BuilderPatternGenerator` to handle the common builder pattern precisely.
Unfortunately, the way we handle it leads to false negatives, since we never collapse propagated taint for all methods returning `this`.
See for example D47437843 which demonstrates a false negative.
This diff simply removes the `no-collapse-on-propagation` mode on methods returning this.
This should not be necessary anymore since we properly infer the tito collapse depth.

Reviewed By: yuhshin-oss

Differential Revision: D47438059

fbshipit-source-id: fa9edf0b0ae13df1fc5dbeefb749fbed0e78e4a0