FramesMap: Use HashedAbstractPartition instead of PatriciaTreeMapAbstractPartition
Summary:
PatriciaTreeMap does not handle key collisions. There should be no collisions if the key is a pointer (e.g. `Positions*`), but if we are to use this for other `*Frames` classes, keys like `AccessPath` will need a crypto hash to ensure uniqueness. Computing this might be expensive since `Taint` is used a lot during the analysis.

Opting for `HashedAbstractPartition` instead which uses `std::unordered_map` internally and handles key collisions. It is missing the `map` and `difference_like_operation` which is implemented within `FramesMap`.

Reviewed By: arthaud

Differential Revision: D47383246

fbshipit-source-id: 601c57441e2d54fdfee77cf86aecf563e333eea0