Add KindFrames (contains callee interval)
Summary:
For class intervals work. `KindFrames` will contain the Interval -> Frame map for a given kind.

`CalleePortFrames` which currently contains a Kind -> Frames map will be updated to use this in a Kind -> KindFrames map.

Currently only stores the interval but doesn't do anything with it. Intersecting and propagating intervals correctly will be done after this.

Reviewed By: anwesht

Differential Revision: D47535610

fbshipit-source-id: d19c0fbed2b42398bbb5101d63449d7b89c6bc81