Test case for propagation of user features
Summary:
When a feature is user-declared (e.g. in model generator), it shows up as "user features" in the `CallInfo::Declaration` frame. However, declaration frames do not show up in the UI. As a result, the user never sees where these features were introduced.

This stack will fix the issue. Meanwhile, this change adds a test case for the scenario.

Reviewed By: anwesht

Differential Revision: D47982621

fbshipit-source-id: 734e964584e199061e5e05ac167dcf0f18cb5069