Emit class intervals in json
Summary:
Emit intervals in the JSON. These are not parsed or used in the UI yet. For now, they are used for (manual) verification in E2E tests.

Only non-default intervals are emitted for now to reduce verbosity of test output. All of the will eventually need to be emitted when class interval trace support is implemented.

Reviewed By: anwesht

Differential Revision: D47803781

fbshipit-source-id: 4a87262ec7882b3628e123ce60f51db519ac86a5