Add new CallInfo for PropagationWithTrace
Summary:
This diff stack implements adding subtraces for taint-transforms.
For this, we need to track the "next hops" of propagations with taint transforms. This diff updates the `CallInfo` for this and adds `PropagationWithTrace` vs a regular `Propagation`. As the Taints of type `PropagationWithTrace` will additionally track the "next hops" for propagations with taint transforms, we also need to know if the *call type* of the next hop is declaration/origin/callsite. This is done using `Flags`. To make sure that the `CallInfo` is always initialized correctly, we have named constructors.

Reviewed By: yuhshin-oss

Differential Revision: D47778560

fbshipit-source-id: 2256e3bedd9ad450e581c88026055f68d4620d38